requirements.txt
requests
google-generativeai
google-auth-oauthlib
